+++ hombreslobo-typescript/index.html
@@ -5,6 +5,10 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Castronegro – Hombres Lobo</title>
     <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&icon_names=logout"
+    />
+    <link
       href="https://fonts.googleapis.com/css2?family=Bahiana&display=swap"
       rel="stylesheet"
     />
