+++ hombreslobo-typescript/src/autenticacion/htmls/login.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Iniciar Sesión – Castronegro</title>
+
     <link
       href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;700&family=Merriweather:wght@700&display=swap"
       rel="stylesheet"
